refactor(datatable): move buildSearchBarLayout to layout/ (refs #298)
